requirements.txt
streamlit>=1.45.0
langchain>=0.3.0
langchain-google-genai>=2.1.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
langchain-text-splitters>=0.3.0
tiktoken>=0.8.0
supabase>=2.0.0
pytest>=7.0.0
